No longer crashes with errors, but Xargon halts somewhere mysterious now.

+++ include/music.h
@@ -1,17 +1,19 @@
 //		MUSIC.H Generalized sound driver - header
 //		Copyright 1991 Epic MegaGames, Written by Tim Sweeney
 
+#include <stdint.h>
+
 #define num_samps 50
 #define num_text  40
 #define num_macs  128
 #define maxsndlen 4096
 
-extern long vocposn [num_samps];
-extern int voclen [num_samps];
-extern int vocrate [num_samps];
+extern int32_t vocposn [num_samps];
+extern int16_t voclen [num_samps];
+extern int16_t vocrate [num_samps];
 extern char *soundmac [num_macs];
-extern long textposn [num_text];
-extern int textlen [num_text];
+extern int32_t textposn [num_text];
+extern int16_t textlen [num_text];
 
 extern int *freq;
 extern int *dur;
